pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "atoevidence"
version = "1.1.0"
description = "ATO evidence traceability automation (portfolio reference)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = ["PyYAML>=6.0", "click>=8.1"]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
atoevidence = "atoevidence.cli:main"
